chore: reorganize utils
